+++ changelog.md
@@ -5,6 +5,7 @@
 - Fixed bug with accounting for HVAC delivered heat for standalone HVAC runs 
 - Fixed bug with ASHP backup heater units
 - Added OCHREException class to handle errors
+- Fixed bug with state space model reduction algorithm
 
 ### OCHRE v0.8.4-beta